requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2.7
django-autoslug==1.9.9
django-autoslug-field==0.2.3
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.2.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.9
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.13.2
Unidecode==1.3.8
urllib3==2.4.0
